sort the list, Namirah Commit20

+++ src/PizzaBar/CompareByTime.java
@@ -8,6 +8,6 @@ public class CompareByTime implements Comparator<Order> {
 
     @Override
     public int compare(Order o1, Order o2){
-        return Double.compare(o2.minutesToPickup(), o1.minutesToPickup());
+        return Long.compare(o2.minutesToPickup(), o1.minutesToPickup());
     }
 }
